+++ src/CruiseControl.h
@@ -26,8 +26,8 @@ private:
     bool _isEnabled = false;
     bool _needToGetSensorsValue = false;
     bool _isSpeedControl = false;
-    int16_t _targetSpeed = 0;
-    int16_t _lastTargetSpeed = 0;
+    double _targetSpeed = 0;
+    double _lastTargetSpeed = 0;
     int _pedal0 = 0;
     int _pedal1 = 0;
     int _pedal0min = 0;
